% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_entropy1.R
\name{conf_entropy1}
\alias{conf_entropy1}
\title{Configurational entropy for landscape mosaics}
\usage{
conf_entropy1(x, nr_of_permutations, independent = FALSE)
}
\arguments{
\item{x}{SpatRaster object (\code{\link[terra:rast]{terra::rast()}}) containing one or more categorical rasters}

\item{nr_of_permutations}{Number of permutations performed on each input raster to calculate
possible distribution of total edge values}

\item{independent}{Should an independent set of permutations be performed for each input raster?
\code{TRUE}/\code{FALSE}.
Use \code{FALSE} (default) when each of your input rasters has the same configuration.}
}
\value{
A tibble
}
\description{
Configurational entropy for landscape mosaics
}
\examples{
library(terra)
library(bespatial)
mosaic = rast(system.file("raster/mosaic.tif", package = "bespatial"))
ce1 = conf_entropy1(mosaic, 1000)
plot(ce1$total_edge, ce1$lnp)
conf_entropy1(mosaic, 1000, independent = TRUE)
}
