% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_entropy2.R
\name{conf_entropy2}
\alias{conf_entropy2}
\title{Configurational entropy for surfaces}
\usage{
conf_entropy2(x, nr_of_permutations, independent = FALSE)
}
\arguments{
\item{x}{SpatRaster object (\code{\link[terra:rast]{terra::rast()}}) containing one or more continuous rasters}

\item{nr_of_permutations}{Number of permutations performed on each input raster to calculate
possible distribution of "slope" values}

\item{independent}{Should an independent set of permutations be performed for each input raster?
\code{TRUE}/\code{FALSE}.
Use \code{FALSE} (default) when each of your input rasters has the same configuration.}
}
\value{
A tibble
}
\description{
Configurational entropy for surfaces
}
\examples{
library(terra)
library(bespatial)
surface = rast(system.file("raster/surface.tif", package = "bespatial"))
ce2 = conf_entropy2(surface, 1000)
plot(ce2$slope, ce2$lnp)
conf_entropy2(surface, 1000, independent = TRUE)
}
