% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_entropy3.R
\name{conf_entropy3}
\alias{conf_entropy3}
\title{Configurational entropy for point patterns}
\usage{
conf_entropy3(x, nr_of_permutations, independent = FALSE)
}
\arguments{
\item{x}{SpatRaster object (\code{\link[terra:rast]{terra::rast()}}) containing one or more rasters with one value and NAs}

\item{nr_of_permutations}{Number of permutations performed on each input raster to calculate
possible distribution of the number of nearest neighbors}

\item{independent}{Should an independent set of permutations be performed for each input raster?
\code{TRUE}/\code{FALSE}.
Use \code{FALSE} (default) when each of your input rasters has the same configuration.}
}
\value{
A tibble
}
\description{
Configurational entropy for point patterns
}
\examples{
library(terra)
library(bespatial)
point_pattern = rast(system.file("raster/point_pattern.tif", package = "bespatial"))
ce3 = conf_entropy3(point_pattern, 100)
plot(ce3$distance, ce3$lnp)
ce3b = conf_entropy3(point_pattern, 100, independent = TRUE)
plot(ce3b$distance, ce3b$lnp)
}
