% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Calculates an average distance between non-NA cells}
\usage{
get_distance(p, x)
}
\arguments{
\item{p}{A matrix}

\item{x}{A SpatRaster with proper metadata (e.g., extent and CRS)}
}
\value{
An average distance between points
}
\description{
Calculates an average distance between non-NA cells
}
\details{
It converts permuted matrix into a vector dataset, and
calculates an average distance between the points
}
