% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_edge.R
\name{get_total_edge}
\alias{get_total_edge}
\title{Calculate total edge based on the input matrix}
\usage{
get_total_edge(x, resolution, neighbourhood = as.matrix(4))
}
\arguments{
\item{x}{A matrix}

\item{resolution}{A numeric vector with two values representing the input matrix
resolution on the x and y axis}

\item{neighbourhood}{The number of directions in which cell adjacencies are
considered as neighbours: 4 (rook's case), 8 (queen's case) or a binary matrix
where the ones define the neighbourhood. The default is 4.}
}
\value{
A total edge value
}
\description{
Calculate total edge based on the input matrix
}
