% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_raster.R
\name{permute_raster}
\alias{permute_raster}
\title{Permute values in the input raster}
\usage{
permute_raster(x, nr_of_permutations)
}
\arguments{
\item{x}{SpatRaster object (\code{\link[terra:rast]{terra::rast()}}) containing one or more rasters}

\item{nr_of_permutations}{Number of permutations performed on each input raster}
}
\value{
A list of matrices
}
\description{
Permute values in the input raster
}
